feat #24: friend search revalidation
